Revamp load position / games feature
