+++ lib/src/view/board_editor/board_editor_screen.dart
@@ -2,6 +2,7 @@ import 'package:chessground/chessground.dart';
 import 'package:dartchess/dartchess.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:lichess_mobile/src/constants.dart';
 import 'package:lichess_mobile/src/model/analysis/analysis_controller.dart';
@@ -17,6 +18,7 @@ import 'package:lichess_mobile/src/utils/screen.dart';
 import 'package:lichess_mobile/src/utils/share.dart';
 import 'package:lichess_mobile/src/view/analysis/analysis_screen.dart';
 import 'package:lichess_mobile/src/view/board_editor/board_editor_filters.dart';
+import 'package:lichess_mobile/src/view/board_editor/board_editor_positions.dart';
 import 'package:lichess_mobile/src/view/play/create_challenge_bottom_sheet.dart';
 import 'package:lichess_mobile/src/view/user/search_screen.dart';
 import 'package:lichess_mobile/src/widgets/adaptive_action_sheet.dart';
@@ -43,6 +45,17 @@ class BoardEditorScreen extends ConsumerWidget {
       appBar: AppBar(
         title: Text(context.l10n.boardEditor),
         actions: [
+          IconButton(
+            icon: const Icon(Icons.edit),
+            tooltip: 'FEN',
+            onPressed: () => showDialog<void>(
+              context: context,
+              builder: (_) => _FenDialog(
+                onFenLoaded: (fen) =>
+                    ref.read(boardEditorControllerProvider(initialFen).notifier).loadFen(fen),
+              ),
+            ),
+          ),
           SemanticIconButton(
             semanticsLabel: context.l10n.mobileSharePositionAsFEN,
             onPressed: () => launchShareDialog(context, ShareParams(text: boardEditorState.fen)),
@@ -284,7 +297,7 @@ class _BottomBar extends ConsumerWidget {
                 onPressed: () {
                   final notifier = ref.read(editorController.notifier);
                   Navigator.of(context).push(
-                    SearchPositionScreen.buildRoute(
+                    BoardEditorPositionsScreen.buildRoute(
                       context,
                       onPositionSelected: (position) => {
                         notifier.loadFen(position.fen),
@@ -384,6 +397,61 @@ class _BottomBar extends ConsumerWidget {
           icon: Icons.tune,
         ),
       ],
+    );
+  }
+}
+
+class _FenDialog extends StatefulWidget {
+  const _FenDialog({required this.onFenLoaded});
+
+  final void Function(String fen) onFenLoaded;
+
+  @override
+  State<_FenDialog> createState() => _FenDialogState();
+}
+
+class _FenDialogState extends State<_FenDialog> {
+  final _controller = TextEditingController();
+
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
+
+  Future<void> _pasteFromClipboard() async {
+    final ClipboardData? data = await Clipboard.getData(Clipboard.kTextPlain);
+    if (data?.text == null || !mounted) return;
+
+    final text = data!.text!.trim();
+    if (text.isEmpty) return;
+
+    _controller.text = text;
+    Navigator.of(context, rootNavigator: true).pop();
+    try {
+      final pos = Chess.fromSetup(Setup.parseFen(text));
+      widget.onFenLoaded(pos.fen);
+    } catch (_) {
+      showSnackBar(context, context.l10n.invalidFen, type: SnackBarType.error);
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      content: TextField(
+        controller: _controller,
+        readOnly: true,
+        onTap: _pasteFromClipboard,
+        decoration: InputDecoration(
+          hintText: context.l10n.pasteTheFenStringHere,
+          suffixIcon: IconButton(
+            icon: const Icon(Icons.paste),
+            onPressed: _pasteFromClipboard,
+            tooltip: 'Paste from clipboard',
+          ),
+        ),
+      ),
     );
   }
 }
